# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InstaApp API",
        "description": "API backend untuk aplikasi InstaApp",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/follow/{id}": {
            "post": {
                "tags": [
                    "Follow"
                ],
                "summary": "Follow a user",
                "description": "Follow User.",
                "operationId": "456efed9e5384734c6c9e8bb338d8672",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of user to follow",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Followed"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Follow"
                ],
                "summary": "Unfollow a user",
                "description": "Unfollow User.",
                "operationId": "501c1186465a45abc03f59cb540dd31a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of user to unfollow",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unfollowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "operationId": "29655023c0605f4d72de82f9147e3c5b",
                "responses": {
                    "200": {
                        "description": "List of users"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get authenticated user info",
                "operationId": "ea32fa4cf836c7fa82b9db1778c97c5e",
                "responses": {
                    "200": {
                        "description": "Authenticated user data"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get detail of a user",
                "operationId": "e19f5cd7ea87abe5ba0951304de94f7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail user"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile": {
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update user profile",
                "operationId": "05832ea3ee2a2beb2343fb2bac2eee80",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "bio": {
                                        "type": "string",
                                        "example": "This is my bio."
                                    },
                                    "website": {
                                        "type": "string",
                                        "example": "https://example.com"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "/uploads/avatar.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/{id}": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get user profile detail",
                "operationId": "e59fdcd1e7afca9a2ef91afbba373afe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile fetched"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}